moving scripts for processing ONS population data for QA purposes
